+++ tests/test_download_photos.py
@@ -2389,3 +2389,88 @@ class DownloadPhotoTestCase(TestCase):
             self._caplog.text,
         )
         self.assertIn("INFO     All photos have been downloaded", self._caplog.text)
+
+    def test_download_and_skip_old(self) -> None:
+        base_dir = os.path.join(self.fixtures_path, inspect.stack()[0][3])
+
+        files_to_create = [
+            # ("2018/07/30", "IMG_7408.JPG", 1151066),
+            # ("2018/07/30", "IMG_7407.JPG", 656257),
+        ]
+
+        files_to_download = [("2018/07/31", "IMG_7409.JPG")]
+
+        data_dir, result = run_icloudpd_test(
+            self.assertEqual,
+            self.root_path,
+            base_dir,
+            "listing_photos.yml",
+            files_to_create,
+            files_to_download,
+            [
+                "--username",
+                "jdoe@gmail.com",
+                "--password",
+                "password1",
+                "--recent",
+                "5",
+                "--skip-videos",
+                "--skip-live-photos",
+                "--set-exif-datetime",
+                "--no-progress-bar",
+                "--skip-created-before",
+                "2018-07-31",
+            ],
+        )
+
+        assert result.exit_code == 0
+
+        self.assertIn("DEBUG    Looking up all photos...", self._caplog.text)
+        self.assertIn(
+            f"INFO     Downloading 5 original photos to {data_dir} ...",
+            self._caplog.text,
+        )
+        for dir_name, file_name in files_to_download:
+            file_path = os.path.normpath(os.path.join(dir_name, file_name))
+            self.assertIn(
+                f"DEBUG    Downloading {os.path.join(data_dir, file_path)}",
+                self._caplog.text,
+            )
+        self.assertNotIn(
+            "IMG_7409.MOV",
+            self._caplog.text,
+        )
+        for dir_name, file_name in [
+            (dir_name, file_name) for (dir_name, file_name, _) in files_to_create
+        ]:
+            file_path = os.path.normpath(os.path.join(dir_name, file_name))
+            self.assertIn(
+                f"DEBUG    {os.path.join(data_dir, file_path)} already exists",
+                self._caplog.text,
+            )
+
+        self.assertIn(
+            "DEBUG    Skipping IMG_7405.MOV, only downloading photos.",
+            self._caplog.text,
+        )
+        self.assertIn(
+            "DEBUG    Skipping IMG_7404.MOV, only downloading photos.",
+            self._caplog.text,
+        )
+        self.assertIn(
+            "DEBUG    Skipping IMG_7407.JPG, as it was created 2018-07-30 11:44:05.108000+00:00, before 2018-07-31 00:00:00+00:00.",
+            self._caplog.text,
+        )
+        self.assertIn(
+            "DEBUG    Skipping IMG_7408.JPG, as it was created 2018-07-30 11:44:10.176000+00:00, before 2018-07-31 00:00:00+00:00.",
+            self._caplog.text,
+        )
+        self.assertIn("INFO     All photos have been downloaded", self._caplog.text)
+
+        # Check that file was downloaded
+        # Check that mtime was updated to the photo creation date
+        photo_mtime = os.path.getmtime(
+            os.path.join(data_dir, os.path.normpath("2018/07/31/IMG_7409.JPG"))
+        )
+        photo_modified_time = datetime.datetime.fromtimestamp(photo_mtime, datetime.timezone.utc)
+        self.assertEqual("2018-07-31 07:22:24", photo_modified_time.strftime("%Y-%m-%d %H:%M:%S"))